+Integrated player classes. +Added player to game. +Refactored project structure and some code. +Added InputManager.
-Deleted unnecessary code.

+++ AlkonostXNA/AlkonostXNA/AlkonostDataStructure/Data/Axe.cs
@@ -7,10 +7,10 @@ namespace AlkonostXNAGame.AlkonostDataStructure.Data
 {
     public class Axe : Item
     {
-        private const double DefautMovementBonus = 0;
-        private const double DefautDamage = 0;
-        private const double DefautArmor = 30;
-        private const double DefautHealth = 0;
+        private const float DefautMovementBonus = 0;
+        private const float DefautDamage = 0;
+        private const float DefautArmor = 30;
+        private const float DefautHealth = 0;
         private const int DefautCritChannce = 1;
 
         public Axe() : base(DefautDamage, DefautArmor, DefautHealth, DefautMovementBonus, DefautCritChannce)
