+++ AlkonostXNA/AlkonostXNA/AlkonostDataStructure/Data/Boots.cs
@@ -7,10 +7,10 @@ namespace AlkonostXNAGame.AlkonostDataStructure.Data
 {
     public class Boots : Item
     {
-        private const double DefautMovementBonus = 60;
-        private const double DefautDamage = 0;      
-        private const double DefautArmor = 0;
-        private const double DefautHealth = 0;
+        private const float DefautMovementBonus = 60;
+        private const float DefautDamage = 0;      
+        private const float DefautArmor = 0;
+        private const float DefautHealth = 0;
         private const int DefautCritChannce = 0;
 
         public Boots() : base(DefautDamage, DefautArmor, DefautHealth, DefautMovementBonus, DefautCritChannce)
