+++ AlkonostXNA/AlkonostXNA/AlkonostDataStructure/Data/Flail.cs
@@ -7,10 +7,10 @@ namespace AlkonostXNAGame.AlkonostDataStructure.Data
 {
     public class Flail : Item
     {
-        private const double DefautMovementBonus = 0;
-        private const double DefautDamage = 40;
-        private const double DefautArmor = 0;
-        private const double DefautHealth = 0;
+        private const float DefautMovementBonus = 0;
+        private const float DefautDamage = 40;
+        private const float DefautArmor = 0;
+        private const float DefautHealth = 0;
         private const int DefautCritChannce = 0;
 
         public Flail() : base(DefautDamage, DefautArmor, DefautHealth, DefautMovementBonus, DefautCritChannce)
