+++ AlkonostXNA/AlkonostXNA/AlkonostDataStructure/Data/Shield.cs
@@ -7,10 +7,10 @@ namespace AlkonostXNAGame.AlkonostDataStructure.Data
 {
     public class Shield : Item
     {
-        private const double DefautMovementBonus = 0;
-        private const double DefautDamage = 0;
-        private const double DefautArmor = 40;
-        private const double DefautHealth = 50;
+        private const float DefautMovementBonus = 0;
+        private const float DefautDamage = 0;
+        private const float DefautArmor = 40;
+        private const float DefautHealth = 50;
         private const int DefautCritChannce = 0;
 
         public Shield() : base(DefautDamage, DefautArmor, DefautHealth, DefautMovementBonus, DefautCritChannce)
